fixing refactoring issues
