+++ src/actions/index.js
@@ -1,5 +1,27 @@
-export { ADD_USER_START, ADD_USER_FAIL, ADD_USER_SUCCESS } from './addUser';
+export {
+  ADD_USER_START,
+  ADD_USER_FAIL,
+  ADD_USER_SUCCESS,
+  addUser,
+} from './addUser';
 
-export { SUBMIT_START, SUBMIT_FAIL, SUBMIT_SUCCESS } from './submit.js';
+export {
+  SUBMIT_START,
+  SUBMIT_FAIL,
+  SUBMIT_SUCCESS,
+  getData,
+} from './submit.js';
 
-export { RANDOM_USER, RANDOM_FAIL, RANDOM_SUCCESS } from './randomUsers';
+export {
+  RANDOM_USER,
+  RANDOM_FAIL,
+  RANDOM_SUCCESS,
+  random,
+} from './randomUsers';
+
+export {
+  CREATE_LISTING_FAILED,
+  CREATE_LISTING_START,
+  CREATE_LISTING_SUCCESS,
+  createListing,
+} from './createListing';
