+++ src/components/Login/Login.js
@@ -1,12 +1,10 @@
 import React, { useState } from 'react';
 
 import { useSelector, useDispatch } from 'react-redux';
-import * as actions from '../../actions';
 
 const Login = props => {
-  const auth = useSelector(store => store.auth);
+  const auth = useSelector(store => store.submit.auth);
   const content = (
-    //  conditionally renders content based on login form or sign up form state.
     <div className="App">
       <h1>Tech2Rent</h1>
       <button onClick={auth.login}>Login/Register</button>
